Update blocks.py
The pddl description of a state does not take into account the location of the different towers of blocks (or of the single blocks) on the table, e.g. "ontable A" means that block A is indifferently on the left, middle, right or any position on the table.

This is not expressed in the render() function yet. So I added a line to simply shuffle randomly the position of locations, so that this indifferentiability is "taken into account". Without this, it's impossible to properly explore the true width of dataset